Refactor the errorMiddleware
- put errorMiddeware in a seperate file
- print out unhandeld rejections
