+++ .gitignore
@@ -11,6 +11,9 @@ pids
 *.seed
 *.pid.lock
 
+# Data
+data/*
+
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
